+++ ReassureTest.Net.Tests/NestedTests.cs
@@ -1,5 +1,3 @@
-using System.Runtime.CompilerServices;
-using System;
 using NUnit.Framework;
 
 namespace ReassureTest.Tests