+++ ReassureTest.Net/AST/Expected/AstDateTimeMatcher.cs
@@ -13,4 +13,14 @@ namespace ReassureTest.AST.Expected
             AcceptedSlack = acceptedSlack;
         }
     }
+
+    class AstStringMatcher : IAssertEvaluator
+    {
+        public AstSimpleValue UnderlyingValue { get; set; }
+        
+        public AstStringMatcher(AstSimpleValue value)
+        {
+            UnderlyingValue = value;
+        }
+    }
 }