+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Version history
 
 
+## v0.5.0
+* Bugfix some datetime compare output
+* Add string match with `*`
+
 ## v0.1.0
 * Handling uninitialized `ImmutableArray<>`
 * Empty objects returns empty results